Create an analog to CSS
An additional style can be added to style prop wrapping the content of it into square brackets, which would build an array of styles.
Then we pass a style from a child component as a prop.
The last added style (on the right) can overwrite the style standing on the left in an array